+++ pkgs/gtest.ps1
@@ -1,7 +1,8 @@
 #Requires -RunAsAdministrator
 
 $ErrorActionPreference="Stop"
-& "$(Split-Path -Path $MyInvocation.MyCommand.Path -Parent)/env/mirror.ps1"
+& "$(Split-Path -Path $MyInvocation.MyCommand.Path -Parent)/env/mirror.ps1" | Out-Null
+& "$(Split-Path -Path $MyInvocation.MyCommand.Path -Parent)/env/toolchain.ps1" | Out-Null
 
 pushd ${Env:TMP}
 $repo="${Env:GIT_MIRROR}/google/googletest.git"
@@ -27,32 +28,35 @@ pushd build
 
 $gtest_silent_warning="/D_SILENCE_STDEXT_HASH_DEPRECATION_WARNINGS /D_SILENCE_TR1_NAMESPACE_DEPRECATION_WARNING /w"
 
+# /Zi is hard-coded in googletest/cmake/internal_utils.cmake.
 cmake                                                               `
-    -DBUILD_GMOCK=ON                                                `
-    -DBUILD_GTEST=ON                                                `
     -DBUILD_SHARED_LIBS=ON                                          `
     -DCMAKE_BUILD_TYPE=Release                                      `
-    -DCMAKE_C_FLAGS="/GL /MP /Z7"                                   `
-    -DCMAKE_CXX_FLAGS="/EHsc /GL /MP /Z7 ${gtest_silent_warning}"   `
-    -DCMAKE_EXE_LINKER_FLAGS="/LTCG:incremental"                    `
-    -DCMAKE_EXE_LINKER_FLAGS_RELWITHDEBINFO="/INCREMENTAL:NO"       `
+    -DCMAKE_C_FLAGS="/GL /MP"                                       `
+    -DCMAKE_CXX_FLAGS="/EHsc /GL /MP ${gtest_silent_warning}"       `
+    -DCMAKE_EXE_LINKER_FLAGS="/DEBUG:FASTLINK /LTCG:incremental"    `
     -DCMAKE_INSTALL_PREFIX="${Env:ProgramFiles}/googletest"         `
-    -DCMAKE_SHARED_LINKER_FLAGS="/LTCG:incremental"                 `
-    -DCMAKE_SHARED_LINKER_FLAGS_RELWITHDEBINFO="/INCREMENTAL:NO"    `
+    -DCMAKE_SHARED_LINKER_FLAGS="/DEBUG:FASTLINK /LTCG:incremental" `
     -DCMAKE_STATIC_LINKER_FLAGS="/LTCG:incremental"                 `
     -Dgmock_build_tests=OFF                                         `
     -Dgtest_build_samples=ON                                        `
     -Dgtest_build_tests=OFF                                         `
-    -G"Visual Studio 15 2017 Win64"                                 `
-    -T"host=x64"                                                    `
+    -G"Ninja"                                                       `
     ..
 
-cmake --build . --config Release -- -maxcpucount
+cmake --build .
 
-cmake --build . --config Release --target run_tests -- -maxcpucount
+$ErrorActionPreference="SilentlyContinue"
+cmake --build . --target test
+if (-Not $?)
+{
+    echo "Oops! Expect to pass all tests."
+    exit 1
+}
+$ErrorActionPreference="Stop"
 
 rm -Force -Recurse -ErrorAction SilentlyContinue -WarningAction SilentlyContinue "${Env:ProgramFiles}/googletest"
-cmake --build . --config Release --target install -- -maxcpucount
+cmake --build . --target install
 Get-ChildItem "${Env:ProgramFiles}/googletest" -Filter *.dll -Recurse | Foreach-Object { New-Item -Force -ItemType SymbolicLink -Path "${Env:SystemRoot}\System32\$_" -Value $_.FullName }
 
 # Alias to default names in CMake.
