+++ pkgs/snappy.ps1
@@ -1,7 +1,8 @@
 #Requires -RunAsAdministrator
 
 $ErrorActionPreference="Stop"
-& "$(Split-Path -Path $MyInvocation.MyCommand.Path -Parent)/env/mirror.ps1"
+& "$(Split-Path -Path $MyInvocation.MyCommand.Path -Parent)/env/mirror.ps1" | Out-Null
+& "$(Split-Path -Path $MyInvocation.MyCommand.Path -Parent)/env/toolchain.ps1" | Out-Null
 
 pushd ${Env:TMP}
 $repo="${Env:GIT_MIRROR}/google/snappy.git"
@@ -16,33 +17,39 @@ if (Test-Path "$root")
 }
 
 $latest_ver=$($(git ls-remote --tags "$repo") -match '.*refs/tags/[0-9\.]*$' -replace '.*refs/tags/','' | sort {[Version]$_})[-1]
+$latest_ver="master"
 git clone --depth 1 --single-branch -b "$latest_ver" "$repo"
 pushd "$root"
 mkdir build
 pushd build
 
+# /GL doesn't work for unknown reason.
 cmake                                                               `
-    -A x64                                                          `
     -DBUILD_SHARED_LIBS=ON                                          `
-    -DCMAKE_BUILD_TYPE=RelWithDebInfo                               `
-    -DCMAKE_C_FLAGS="/MP"                                           `
-    -DCMAKE_CXX_FLAGS="/EHsc /MP"                                   `
-    -DCMAKE_EXE_LINKER_FLAGS="/LTCG:incremental"                    `
-    -DCMAKE_EXE_LINKER_FLAGS_RELWITHDEBINFO="/INCREMENTAL:NO"       `
+    -DCMAKE_BUILD_TYPE=Release                                      `
+    -DCMAKE_C_FLAGS="/MP /Z7"                                       `
+    -DCMAKE_CXX_FLAGS="/EHsc /MP /Z7"                               `
+    -DCMAKE_EXE_LINKER_FLAGS="/DEBUG:FASTLINK /LTCG:incremental"    `
     -DCMAKE_INSTALL_PREFIX="${Env:ProgramFiles}/Snappy"             `
-    -DCMAKE_SHARED_LINKER_FLAGS="/LTCG:incremental"                 `
-    -DCMAKE_SHARED_LINKER_FLAGS_RELWITHDEBINFO="/INCREMENTAL:NO"    `
+    -DCMAKE_SHARED_LINKER_FLAGS="/DEBUG:FASTLINK /LTCG:incremental" `
     -DCMAKE_STATIC_LINKER_FLAGS="/LTCG:incremental"                 `
-    -G"Visual Studio 15 2017"                                       `
-    -T"host=x64"                                                    `
+    -DSNAPPY_REQUIRE_AVX2=ON                                        `
+    -G"Ninja"                                                       `
     ..
 
-cmake --build . --config RelWithDebInfo -- -maxcpucount
+cmake --build .
 
-cmake --build . --config RelWithDebInfo --target run_tests -- -maxcpucount
+$ErrorActionPreference="SilentlyContinue"
+cmake --build . --config run_tests
+if (-Not $?)
+{
+    echo "Oops! Expect to pass all tests."
+    exit 1
+}
+$ErrorActionPreference="Stop"
 
 rm -Force -Recurse -ErrorAction SilentlyContinue -WarningAction SilentlyContinue "${Env:ProgramFiles}/Snappy"
-cmake --build . --config RelWithDebInfo --target install -- -maxcpucount
+cmake --build . --target install
 Get-ChildItem "${Env:ProgramFiles}/Snappy" -Filter *.dll -Recurse | Foreach-Object { New-Item -Force -ItemType SymbolicLink -Path "${Env:SystemRoot}\System32\$_" -Value $_.FullName }
 
 popd
