+++ pkgs/zlib.ps1
@@ -1,7 +1,8 @@
 #Requires -RunAsAdministrator
 
 $ErrorActionPreference="Stop"
-& "$(Split-Path -Path $MyInvocation.MyCommand.Path -Parent)/env/mirror.ps1"
+& "$(Split-Path -Path $MyInvocation.MyCommand.Path -Parent)/env/mirror.ps1" | Out-Null
+& "$(Split-Path -Path $MyInvocation.MyCommand.Path -Parent)/env/toolchain.ps1" | Out-Null
 
 pushd ${Env:TMP}
 $repo="${Env:GIT_MIRROR}/madler/zlib.git"
@@ -23,26 +24,30 @@ mkdir build
 pushd build
 
 # Currently (v1.2.11) there's a bug related to AMD64 flag.
-cmake                                                               `
-    -A x64                                                          `
-    -DAMD64=OFF                                                     `
-    -DCMAKE_BUILD_TYPE=RelWithDebInfo                               `
-    -DCMAKE_C_FLAGS="/GL /MP /guard:cf"                             `
-    -DCMAKE_EXE_LINKER_FLAGS="/LTCG:incremental /guard:cf"          `
-    -DCMAKE_EXE_LINKER_FLAGS_RELWITHDEBINFO="/INCREMENTAL:NO"       `
-    -DCMAKE_INSTALL_PREFIX="${Env:ProgramFiles}/zlib"               `
-    -DCMAKE_SHARED_LINKER_FLAGS="/LTCG:incremental /guard:cf"       `
-    -DCMAKE_SHARED_LINKER_FLAGS_RELWITHDEBINFO="/INCREMENTAL:NO"    `
-    -DCMAKE_STATIC_LINKER_FLAGS="/LTCG:incremental"                 `
-    -G"Visual Studio 15 2017"                                       `
-    -T"host=x64"                                                    `
+cmake                                                                           `
+    -DAMD64=OFF                                                                 `
+    -DCMAKE_BUILD_TYPE=Release                                                  `
+    -DCMAKE_C_FLAGS="/GL /MP /Z7 /guard:cf"                                     `
+    -DCMAKE_EXE_LINKER_FLAGS="/DEBUG:FASTLINK /LTCG:incremental /guard:cf"      `
+    -DCMAKE_INSTALL_PREFIX="${Env:ProgramFiles}/zlib"                           `
+    -DCMAKE_SHARED_LINKER_FLAGS="/DEBUG:FASTLINK /LTCG:incremental /guard:cf"   `
+    -DCMAKE_STATIC_LINKER_FLAGS="/LTCG:incremental"                             `
+    -G"Ninja"                                                                   `
     ..
 
-cmake --build . --config RelWithDebInfo -- -maxcpucount
-cmake --build . --config RelWithDebInfo --target run_tests -- -maxcpucount
+cmake --build .
+
+$ErrorActionPreference="SilentlyContinue"
+cmake --build . --config run_tests
+if (-Not $?)
+{
+    echo "Oops! Expect to pass all tests."
+    exit 1
+}
+$ErrorActionPreference="Stop"
 
 rm -Force -Recurse -ErrorAction SilentlyContinue -WarningAction SilentlyContinue "${Env:ProgramFiles}/zlib"
-cmake --build . --config RelWithDebInfo --target install -- -maxcpucount
+cmake --build . --target install
 Get-ChildItem "${Env:ProgramFiles}/zlib" -Filter *.dll -Recurse | Foreach-Object { New-Item -Force -ItemType SymbolicLink -Path "${Env:SystemRoot}\System32\$_" -Value $_.FullName }
 
 popd
